Bug 1577766 - Move render target code into its own file. r=gw
Differential Revision: https://phabricator.services.mozilla.com/D44225

[wrupdater] From https://hg.mozilla.org/mozilla-central/rev/931619f1fe4bd430eb96d911fba70a38341d5646
