+++ webrender/src/renderer.rs
@@ -78,16 +78,17 @@ use rayon::{ThreadPool, ThreadPoolBuilder};
 use crate::record::ApiRecordingReceiver;
 use crate::render_backend::{FrameId, RenderBackend};
 use crate::render_task_graph::RenderTaskGraph;
-use crate::render_task::{RenderTargetKind, RenderTask, RenderTaskData, RenderTaskKind};
+use crate::render_task::{RenderTask, RenderTaskData, RenderTaskKind};
 use crate::resource_cache::ResourceCache;
 use crate::scene_builder::{SceneBuilder, LowPrioritySceneBuilder};
 use crate::screen_capture::AsyncScreenshotGrabber;
 use crate::shade::{Shaders, WrShaders};
 use smallvec::SmallVec;
 use crate::texture_cache::TextureCache;
-use crate::render_task_graph::{AlphaRenderTarget, ColorRenderTarget, PictureCacheTarget};
-use crate::render_task_graph::{BlitJob, BlitJobSource, RenderPassKind, RenderTargetList};
-use crate::render_task_graph::{RenderTarget, TextureCacheRenderTarget};
+use crate::render_target::{AlphaRenderTarget, ColorRenderTarget, PictureCacheTarget};
+use crate::render_target::{RenderTarget, TextureCacheRenderTarget, RenderTargetList};
+use crate::render_target::{RenderTargetKind, BlitJob, BlitJobSource};
+use crate::render_task_graph::RenderPassKind;
 use crate::util::drain_filter;
 
 use std;
